Refactor: Consolidate expression simplification code in `simplify_expression.rs` (#1374)
* Move ConstEvaluator into simplity_expressions.rs

* Move Simplifier into simplify_expressions.rs

* clean up test names and docstrings
